HelloAgent/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "hello-agent"
version = "0.1.0"
description = "A simple AgentCore example agent"
requires-python = ">=3.10"
dependencies = [
    "strands-agents>=1.13.0",
    "strands-agents-tools>=0.2.16",
    "bedrock-agentcore>=1.0.3",
    "python-dotenv>=1.2.1",
    "boto3>=1.42.1",
]

[tool.setuptools.packages.find]
where = ["."]
